requirements.txt
absl-py==1.2.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
chess==1.9.3
colorama==0.4.5
cycler==0.11.0
flatbuffers==2.0.7
fonttools==4.37.1
gast==0.5.3
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
idna==3.4
importlib-metadata==4.12.0
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.23.3
oauthlib==3.2.1
onnxruntime-gpu==1.8.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post2
pygame==2.1.2
pyparsing==3.0.9
pyperclip==1.8.2
python-chess==1.999
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
